server/requirements.txt
aleph_alpha_client==2.16.1
anthropic==0.2.3
cachetools==5.3.0
click==8.1.3
Flask==2.2.3
Flask_Cors==3.0.10
huggingface_hub==0.13.2
openai==0.27.2
psutil==5.9.4
python-dotenv==1.0.0
requests==2.28.2
sentencepiece==0.1.97
six==1.16.0
sseclient==0.0.27
torch==2.0.0
transformers==4.27.1
